chocolate/requirements.txt
tornado
petname
wheel
sqlalchemy
sqlalchemy-migrate
psycopg2